Update README.md
# polkadot-js.github.io

Simple landing and project overview page for the polkadot-js family of repos. It is always accessible via [polkadot.js.org](https://polkadot.js.org/)

## Welcome

This repository provides a collection of tools, utilities, and libraries for interacting with the [Polkadot network](https://polkadot.network) from JavaScript. All code is available on [GitHub](https://github.com/polkadot-js).

## Features

- **User Applications**: User Web and command-line driven applications to perform tasks on the network.
- **Powerful API**: A fully-featured API for interfacing with Polkadot and Substrate chains in your applications.
- **Client**: An alternative implementation of a Polkadot Runtime Environment for substrate chains.

## Layout

While there is a slant towards developer tools, providing libraries to allow others to build tools on top of, a collection of applications are made available that allows for interaction with the network.

- **End-Users**: Interact with [Polkadot/Substrate (latest, project hosted)](https://polkadot.js.org/apps/), [Polkadot (stable, Parity hosted)](https://polkadot.js.org/apps/?rpc=wss://rpc.polkadot.io).
- **Developers**: Get started with the [Polkadot-JS API documentation portal](https://polkadot.js.org/api) for interacting with the network from your apps.

## Contributing

This is a community effort to trial tooling while the network is being developed. Contributions are welcome. Feel free to use the tools and libraries, log issues as you find them, or create pull requests.

- Track issues in the [overall @polkadot project board](https://github.com/orgs/polkadot-js/projects/1).
- Ensure library code is comprehensively [unit tested](https://facebook.github.io/jest/) and interfaces have [full type definitions](https://www.typescriptlang.org/).

## Media

Follow us on [Twitter](https://twitter.com/polkadotjs) to receive the latest updates on progress across the projects.

## License

Apache-2 Licensed | Copyright © 2017-2025 polkadot-js authors and contributors